Major rewrite to slowly move to a scene based setup

+++ DenOfIz.Graphics/Binding/GpuBufferView.cs
@@ -3,6 +3,6 @@ namespace Graphics.Binding;
 public struct GpuBufferView
 {
     public DenOfIz.Buffer Buffer;
-    public ulong NumBytes;
-    public ulong Offset;
+    public uint NumBytes;
+    public uint Offset;
 }
